refactor(backend): fix inverted tasks dependency
process_endpoint_result and get_evaluation_model/get_execution_model
moved out of tasks/ into services/ and utils/, so explorer no longer
imports business logic from the Celery tasks layer. response_extractor.py
moved too, since it's process_endpoint_result's only real dependency and
has no task-specific coupling.
