+++ apps/backend/src/rhesis/backend/app/utils/response_extractor.py
